requirements.txt
fastapi
uvicorn
pyjwt
cryptography
pytest
httpx
